Move authorization to separate file
